Use interpolateUrl in logo src config
