+++ __mocks__/default_config.ts
@@ -1,0 +1,17 @@
+export const mock__defaultConfig = {
+  links: {
+    home: {
+      url: "/home",
+    },
+    account: {
+      url: "/account",
+    },
+    logoutRedirect: {
+      url: "/home",
+    },
+  },
+  brand: {
+    src: null,
+    alt: "OpenMRS",
+  },
+};
