+++ babel.config.json
@@ -6,9 +6,12 @@
   ],
   "plugins": [
     "@babel/plugin-proposal-class-properties",
-    ["i18next-extract", {
-      "outputPath": "translations/{{locale}}.json",
-      "discardOldKeys": true
-    }]
-  ],
+    [
+      "i18next-extract",
+      {
+        "outputPath": "translations/{{locale}}.json",
+        "discardOldKeys": true
+      }
+    ]
+  ]
 }
