+++ src/index.ts
@@ -4,6 +4,7 @@ import {
   getAsyncLifecycle,
   isUrl,
 } from "@openmrs/esm-framework";
+import { configSchema } from "./configSchema";
 
 const backendDependencies = { appui: "^1.0.0" };
 
@@ -15,45 +16,7 @@ const importTranslation = require.context(
 );
 
 function setupOpenMRS() {
-  defineConfigSchema("@pih/esm-refapp-navbar-app", {
-    links: {
-      home: {
-        url: {
-          _type: Type.String,
-          _validators: [isUrl],
-          _default: "${openmrsBase}",
-        },
-      },
-      account: {
-        url: {
-          _type: Type.String,
-          _validators: [isUrl],
-          _default: "${openmrsBase}/adminui/myaccount/myAccount.page",
-        },
-      },
-      logoutRedirect: {
-        url: {
-          _type: Type.String,
-          _validators: [isUrl],
-          _default: "${openmrsBase}",
-        },
-      },
-    },
-    brand: {
-      src: {
-        _type: Type.String,
-        _validators: [isUrl],
-        _default: null,
-        _description:
-          "A path or URL to an image. Defaults to the OpenMRS SVG sprite.",
-      },
-      alt: {
-        _type: Type.String,
-        _default: "OpenMRS",
-        _description: "Alt text, shown if the image fails to load",
-      },
-    },
-  });
+  defineConfigSchema("@pih/esm-refapp-navbar-app", configSchema);
 
   const moduleName = "@pih/esm-refapp-navbar-app";
