+++ tsconfig.json
@@ -16,6 +16,9 @@
       "es2018",
       "es2020"
     ],
+    "paths": {
+      "@openmrs/*": ["./node_modules/@openmrs/*"]
+    },
     "noEmit": true
   }
 }
